Replace text-matching scroll sync with anchor-map interpolation (v2)
V1 used pmPosToMdLine/mdLineToPmPos text matching which drifted

progressively further down the document due to duplicate text and

block-counting mismatches between PM tree and markdown lines.

V2 introduces scroll-map.ts with type-dispatch anchor mapping:

- buildAnchorMap walks PM nodes + markdown lines in parallel

- findNodeInMdLines dispatches on node.type.name, not textContent

- Linear interpolation between anchors for sub-line precision

- Floating-point topLine (e.g. 6.6 = 60% through line 6)

- Cached by PM doc identity to avoid rebuilds on 500ms polls

CodeMirror uses native line API with fractional offset.

Swift CursorPosition.topLine changed from Int to Double.

Updated docs for scroll sync architecture.

+++ final final/Views/ContentView.swift
@@ -13,9 +13,9 @@ struct CursorPosition: Equatable {
     let column: Int
     let scrollFraction: Double
     let cursorIsVisible: Bool
-    let topLine: Int
+    let topLine: Double
 
-    init(line: Int, column: Int, scrollFraction: Double = 0, cursorIsVisible: Bool = true, topLine: Int = 1) {
+    init(line: Int, column: Int, scrollFraction: Double = 0, cursorIsVisible: Bool = true, topLine: Double = 1.0) {
         self.line = line
         self.column = column
         self.scrollFraction = scrollFraction
@@ -23,7 +23,7 @@ struct CursorPosition: Equatable {
         self.topLine = topLine
     }
 
-    static let start = CursorPosition(line: 1, column: 0, scrollFraction: 0, cursorIsVisible: true, topLine: 1)
+    static let start = CursorPosition(line: 1, column: 0, scrollFraction: 0, cursorIsVisible: true, topLine: 1.0)
 }
 
 /// Toast notification shown when entering focus mode, auto-dismisses after 3 seconds
